+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - updated internal ModDevGradle version to 2.0.138
 - updated minimum ModDevGradle version to 2.0.138 for compatibility with unobfuscated sources
+- updated default Java target version to Java 25
 
 ## [1.5.2] - 2026-03-24