+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 - updated default Java target version to Java 25
 - fixed wrong Roughly Enough Items recipe viewer mod id
 - fixed datagen properties not working
+- fixed Maven publishing not using the mod id as path
 
 ## [1.5.2] - 2026-03-24